fix: failed to duplicate version when no jar

+++ HMCL/src/main/java/org/jackhuang/hmcl/game/HMCLGameRepository.java
@@ -129,7 +129,9 @@ public class HMCLGameRepository extends DefaultGameRepository {
 
         if (dstDir.exists()) throw new IOException("Version exists");
         FileUtils.copyDirectory(srcDir.toPath(), dstDir.toPath());
-        Files.move(dstDir.toPath().resolve(srcId + ".jar"), dstDir.toPath().resolve(dstId + ".jar"));
+        if (Files.exists(dstDir.toPath().resolve(srcId + ".jar"))) {
+            Files.move(dstDir.toPath().resolve(srcId + ".jar"), dstDir.toPath().resolve(dstId + ".jar"));
+        }
         Files.move(dstDir.toPath().resolve(srcId + ".json"), dstDir.toPath().resolve(dstId + ".json"));
         VersionSetting oldVersionSetting = getVersionSetting(srcId).clone();
         GameDirectoryType originalGameDirType = oldVersionSetting.getGameDirType();
